9/19/18 Update #1
Created new drop table system. Now instead of extremely linear drops, there are tables for each enemy and can drop multiple different items all with different drop chances and quantities. I've wanted to do this for a long time but it seemed daunting. Not fully implemented yet but the core of it is done.